notify_service/requirements.txt
Django
djangorestframework
requests
dotenvpython-dotenv==0.21.1